refactor(mobile): decompose fizruk PlanCalendar into sub-components
Extract PlanCalendarHeader, CalendarGrid, and DaySheet into
apps/mobile/src/modules/fizruk/components/plan/ and wire them back
into PlanCalendar.tsx. Pure refactor — no behavior, routing, or
prop-contract changes.
